+++ RELEASE_NOTES.md
@@ -1,7 +1,11 @@
 ## Validator Changes
 
-* no changes
+* Add ConceptMap validation
+* Add StructureMap validation
+* Validate using type regex (had been omitted to now, mostly affects decimal validation)
 
 ## Other code changes
 
-* no changes
+* Various fixes and utilities to support StructureMap & ConceptMap validation
+* Fix Observation.value conversion between R5 and other versions
+